[PLAYER-5446] Rename panels to views
