+++ sdk/react/src/views/AdPlaybackScreen/AdBar/AdBar.js
