+++ sdk/react/src/views/AdPlaybackScreen/AdBar/AdBar.styles.json
