+++ sdk/react/src/views/CastConnectedScreen/CastConnectedScreen.js
