+++ sdk/react/src/views/CastConnectedScreen/CastConnectedScreen.styles.json
