+++ sdk/react/src/views/CastConnectedScreen/CastPlayPauseButtons/SwitchButton/SwitchButton.js
