+++ sdk/react/src/views/CastDevicesScreen/CastDeviceListItem/CastDeviceListItem.js
