+++ sdk/react/src/views/DiscoveryPanel/DiscoveryPanel.js
@@ -21,7 +21,7 @@ import ResponsiveList from './ResponsiveList/ResponsiveList';
 import CountdownViewiOS from '../../shared/CountdownTimerIos/CountdownTimerIos';
 import CountdownViewAndroid from '../../shared/CountdownTimerAndroid/CountdownTimerAndroid';
 
-import panelStyles from '../styles/panel.styles';
+import panelStyles from '../styles/view.styles';
 import discoveryPanelStyles from './DiscoveryPanel.styles';
 
 const styles = Utils.getStyles(discoveryPanelStyles);
