+++ sdk/react/src/views/DiscoveryPanel/DiscoveryPanel.styles.json
