+++ sdk/react/src/views/DiscoveryPanel/ResponsiveList/ResponsiveList.js
