+++ sdk/react/src/views/DiscoveryPanel/ResponsiveList/ResponsiveList.styles.json
