+++ sdk/react/src/views/EndScreen/EndScreen.js
