+++ sdk/react/src/views/EndScreen/EndScreen.styles.json
