+++ sdk/react/src/views/EndScreen/InfoPanel/InfoPanel.js
