+++ sdk/react/src/views/EndScreen/InfoPanel/InfoPanel.styles.json
