+++ sdk/react/src/views/ErrorScreen/ErrorScreen.js
