+++ sdk/react/src/views/ErrorScreen/ErrorScreen.styles.json
