+++ sdk/react/src/views/MoreDetailsScreen/MoreDetailsScreen.js
