+++ sdk/react/src/views/MoreDetailsScreen/MoreDetailsScreen.styles.json
