+++ sdk/react/src/views/MoreOptionScreen/MoreOptionScreen.styles.json
