+++ sdk/react/src/views/StartScreen/StartScreen.js
