+++ sdk/react/src/views/StartScreen/StartScreen.styles.json
