+++ sdk/react/src/views/VideoView/UpNext/UpNext.js
