+++ sdk/react/src/views/VideoView/UpNext/UpNext.styles.json
