+++ sdk/react/src/views/VideoView/VideoView.styles.json
