+++ sdk/react/src/views/VideoView/VideoWatermark/VideoWatermark.js
