+++ sdk/react/src/views/VideoView/VideoWatermark/VideoWatermark.styles.json
